Removed fluff and refactored intro.
The logo image from AUTH0 in part suggests that this is the standard authority when they are a 3rd party service provider. The Dilbert cartoon also gives weight to the idea that a JWT session approach addresses security deficiencies of traditional approaches which is not accurate. Beyond that it adds little value other than eye candy. I like eye candy but it needs to be 100% relevant. Started to add in some modified descriptive copy that better reflects my understanding of the value of learning JWT and putting a clearer perspective on why the code in the repo effectively tries to replicate the functionality of traditional session cookies with no clear benefit - the benefit is really only to allow developers to get their hands dirty in manipulating tokens in a simplistic way and should not be used as a default starter platform.